refactor: modularize data pipeline
